requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.11.0
pydantic-settings==2.8.1
pyjwt==2.10.1
python-dotenv==1.1.0
redis==5.2.1
ollama==0.4.7
httpx==0.28.1